% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{plot histogram #' @description function to plot histogram #'}
\usage{
plot_hist(x, breaks, freq)
}
\arguments{
\item{x}{data.frame to plot #' @param breaks a vector giving the breakpoints between histogram cells, as of graphics::hist #' @param freq logical for plotting frequencies, as of graphics::hist #'}
}
\description{
plot histogram #' @description function to plot histogram #'
}
