% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_faithful_hist.R
\name{mod_faithful_hist}
\alias{mod_faithful_hist}
\alias{mod_faithful_hist_ui}
\alias{mod_faithful_hist_server}
\title{Old Faithful histogram Shiny module}
\usage{
mod_faithful_hist_ui(id)

mod_faithful_hist_server(id)
}
\arguments{
\item{id}{Internal parameter for {shiny}.}
}
\description{
UI and server function for an example Shiny module.
}
