% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfaithful-package.R
\docType{package}
\name{shinyfaithful-package}
\alias{shinyfaithful}
\alias{shinyfaithful-package}
\title{shinyfaithful: Shiny4 workshop pkg}
\description{
Package used for the 4th course of the Series Intro to Shiny.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/guidomaggioorg/shinyfaithful}
  \item Report bugs at \url{https://github.com/guidomaggioorg/shinyfaithful/issues}
}

}
\author{
\strong{Maintainer}: Guido Maggio \email{guido.maggio@mirai-solutions.com}

}
\keyword{internal}
